--- app/models/category.ts
@@ -1,23 +1,0 @@
-import { DateTime } from 'luxon'
-import { BaseModel, column } from '@adonisjs/lucid/orm'
-import type { UUID } from 'crypto'
-
-export default class Category extends BaseModel {
-  @column({ isPrimary: true })
-  declare id: number
-
-  @column()
-  declare name:string
-
-  @column()
-  declare isActive:boolean
-
-  @column()
-  declare slug:UUID
-
-  @column.dateTime({ autoCreate: true })
-  declare createdAt: DateTime
-
-  @column.dateTime({ autoCreate: true, autoUpdate: true })
-  declare updatedAt: DateTime
-}
