--- app/models/company.ts
@@ -1,33 +1,0 @@
-import { DateTime } from 'luxon'
-import type { BelongsTo } from '@adonisjs/lucid/types/relations'
-import { BaseModel, belongsTo, column } from '@adonisjs/lucid/orm'
-import User from './user.js'
-
-export default class Company extends BaseModel {
-  @column({ isPrimary: true })
-  declare id: number
-
-  @column()
-  declare slug: string
-
-  @column()
-  declare name: string
-
-  @column()
-  declare description: string
-
-  @column()
-  declare cover: string
-
-  @column()
-  declare userId: BelongsTo<typeof User>
-
-  @column.dateTime({ autoCreate: true })
-  declare createdAt: DateTime
-
-  @column.dateTime({ autoCreate: true, autoUpdate: true })
-  declare updatedAt: DateTime
-
-  @belongsTo(() => User)
-  declare user: BelongsTo<typeof User>
-}
